+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+### 2018-02-27 v2.0.0-rc4
+
+* Added asynchronous query batch processing (QueryAsync<T>) based on IAsyncEnumerable/Ix.NET extensions. Modified from [PR #1](https://github.com/anthonyreilly/NetCoreForce/pull/1)
+* Marked QuerySingle<T> as obsolete: should use standard LINQ Single() instead.
+
 ### 2018-02-27 v2.0.0-rc3
 
 * Added synchronous Username and Password login flow to AuthenticationClient